pyproject.toml
[tool.poetry]
name = "fermerce"
version = "0.1.0"
description = ""
authors = ["kokoserver <kokoserver20@hotmail.com>"]
readme = "README.md"
packages = [{include = "fermerce"}]

[tool.poetry.dependencies]
python = "^3.12"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.3"}
taskiq-aio-pika = "^0.2.0"
kombu = "^5.2.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
deta = "^1.1.0"
taskiq-fastapi = "^0.1.2"
googlemaps = "^4.10.0"
pillow = "^10.4.0"
edgy = {extras = ["test"], version = "^0.13.0"}
uuid7 = "^0.1.0"
esmerald = {extras = ["jwt", "templates", "test"], version = "^3.3.6"}
uvicorn = "^0.30.5"
openpyxl = "^3.1.5"

[tool.poetry.group.test.dependencies]
pytest-asyncio = "^0.20.3"
pytest = "^7.2.2"
faker = "^18.3.1"
pytest-watch = "^4.2.0"

[tool.poetry.group.lint.dependencies]
flake8 = "^6.0.0"


[tool.poetry.group.format.dependencies]
black = "^23.1.0"

[tool.poetry.scripts]
server = 'bootstrap:run_server'


[tool.hatch]
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
